Module 6 Task 1 Код по отрисовке задачи из BoardController перенесен в TaskController
